Align heading terminology and code with the CommonMark standard (#21)
This is a continuation of my standardization. This change breaks up headings into two different types, ATX and setext. This also includes a fix where we were errantly identifying setext headers in the middle of a paragraph (which is invalid).

During a rebase, I also fixed a small typo and import ordering from the thematic breaks change. I'm not too worried about reverting this change and having those changes also get thrown out.