Align link reference definition terminology and logic with CommonMark standard (#24)
This continues my work on aligning the tool better with CommonMark. In this one, I discover a lot of things I didn't account for as I had no idea about titles.

Some of the reformatting should be done in unit tests, but I will do that later. (Famous last words.)