Parsing out sections is now owned completely by the parser
Instead of having a state machine manage what line we're and what we're processing, instead, our new functions take in the entirety of the text we are trying to process, and pop off their portion if the text starts with that section. This gets rid of the need for global variables and will eventually make more of the code readable. This is occuring halfway through standard alignment, so it's a bit dirty, but was necessary to start working on leaf blocks.