Better align list parsing and reformatting with the spec (#57)
This change does introduce a change to how we render Markdown in that we no longer allow indented lists. This was never feature I used and this was the only section type that supported this. The minor version number will be bumped by a future changelist before a new release is pushed to pypi.